requirements.txt
flask
anthropic
openai-whisper
omnilingual-asr
python-dotenv
flask-login
flask-sqlalchemy
google-genai
